Include root removal in the tests
insert() – or the called fixUp() – is not the most computationally
intensive method, both in tests and in the literature; extractRoot() –
that requires fixDown() – is much harder.

For this, root deletion/extraction has been included in the tests: all
items (with random keys) are first inserted and then they are all
removed; in fact, this is very similar to how heapsort works.

It was also required to reduce the number of elements used – both the
default and the actual number used by the author – to compensate for the
added computational cost.

Finally, the new comparison between dheap and jheap (DAryHeap) shows:

 - dheap takes 57% less time than jheap
 - also:
    - dheap uses 75% less instructions than jheap
    - dheap uses 76% less branches than jheap, and there are 67% less
      misspredicted branchs
    - dheap results in 27% less cache misses